pyproject.toml
[tool.poetry]
name = "pythonProjectN1"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.lint.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^0.991"
flake8 = "^7.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '''
/(
    venv
    | .git
    | \.mypy_cache
    | build
    | dist
)/
'''
